Use std::min/std::max

+++ NDNc/pipeline/pipeline-interests-aimd.hpp
@@ -56,8 +56,8 @@ class PipelineInterestsAimd : public PipelineInterests {
     size_t m_windowSize;
     size_t m_windowIncCounter;
     ndn::time::steady_clock::time_point m_lastDecrease;
-    static const unsigned MAX_WINDOW = 65536;
-    static const unsigned MIN_WINDOW = 8;
+    static const size_t MAX_WINDOW = 65536;
+    static const size_t MIN_WINDOW = 8;
     constexpr static const ndn::time::milliseconds MAX_RTT =
         ndn::time::milliseconds{3};
 };
